Fix pgbouncer after refactoring
